requirements.txt
coverage==5.0.3
h5py
numpy
matplotlib==3.1.1
gast==0.2.2
Pillow==7.0.0
seaborn==0.10.0
progress==1.5
tensorflow==2.1.0